requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
spleeter
python-dotenv
yt-dlp